LayoutTests/http/tests/referrer-policy-script/origin/same-origin.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests the behavior of origin referrer policy when same origin.");
jsTestIsAsync = true;

function checkReferrer(value) {
    referrer = value;
    // Should be the origin
    shouldBeEqualToString("referrer", "http://127.0.0.1:8000/");
    finishJSTest();
}
</script>
<script src="http://127.0.0.1:8000/referrer-policy/resources/script.php" referrerpolicy="origin"></script>
</body>
</html>
